Modularize CloudSync credential handling
Extract credential validation, cached prerequisite reads, device identity, and observable block state while preserving the existing auth lifecycle API.
